S2b: migrate the smoke + 5 gas planes to integer Q16.16 semi-Lagrangian (SLint)
Second sub-step of the S2 atmosphere-group fixed-point migration. The smoke field
+ the 5 gas planes (gmap.gas, with gmap.smoke a view of gas[BLACK_SMOKE]) go to
int32 Q16.16; the advection is the INTEGER semi-Lagrangian ("SLint") ported
verbatim from the proven prototype (tools/s2_advection_demo/advection_demo.py,
ceb601b). Determinism-critical: integer +/-/*/>> are exact + order-free, so the
synced smoke/gas state is now bit-identical run-to-run (cross-GPU once S2c lands).
NON-conservative is ACCEPTED (Q-S2-1): the >>16 truncation is a deterministic
gentle decay -> behaviour, not desync; smoke decay is the tuning knob.

NEW SHARED HELPER (fixed_point.h): reciprocal_q16(denom) — the per-cell integer
reciprocal (3-step... 4-step Newton r <- r*(2 - d*r), seeded from a power-of-2
refined to the top two bits, ROUND-TO-NEAREST narrow). Pure integer, GPU-clean.
S2b uses it for the SL bilinear renorm 1/wsum (wsum in (0,1]); S2c reuses it for
the GS Dinv 1/(1+mu*wsum) (denom > 1) — so the seed handles denom > 1 too. The
round-to-nearest narrow is load-bearing: a truncating Newton converges 1 ULP LOW
at the exact points (reciprocal_q16(FP_ONE) would be 65535), which would shave a
uniform 1-ULP decay off every cell at wsum==1.0; rounding makes wsum==1.0 a clean
identity so the only non-conservation is the deliberate sample-truncation decay.
Unit-proved in tests/_s2b_reciprocal_check.cpp (gated by test_s2b_reciprocal.py):
ACCURACY (matches 2^16/denom to <4 counts OR 1e-3 rel over 400k samples incl. the
smallest denominators + both call-site regimes), DETERMINISM, MONOTONICITY (<=1
ULP round wobble), denom<=0 guard.

THE INTEGER-SL PORT (smoke_dynamics.cpp, the SLint scheme):
  * DDA sqrt-free wall-clip march — step the DOMINANT axis cell-by-cell (Chebyshev
    march) toward the source, stop before a sealed tile (no 1-cell tunnel) or ON a
    breach (vent). Replaces the float back-trace-length sqrt. Per-step increment
    is a FLOOR-divide (matches the prototype's Python // for negative upwind disp).
  * Integer bilinear sample — 4 corner-weight Q16.16 products (mul_q16) accumulated
    in int64 (mul_wide, scale 2^32), narrowed once; sealed corners excluded, a
    breach corner contributes 0. No overflow (acc <= 2^34; renorm <= FP_ONE).
  * Renorm 1/wsum via reciprocal_q16 (wsum clamped to FP_ONE>>8 floor), then the
    pinned >>16 (mul_q16) — the gentle non-conservation.
  * Wind-coupled diffusion stays, integer: the permeability-weighted 4-neighbour
    Laplacian (neighbor_q), then smoke += mul_q16(quantize(d_eff*dt), lap).
  * sink_hop stays the SL-style integer breach-pull (Q-S2-5), same machinery.

FLOAT BRIDGES (collapse at S2c / when fire+steam migrate), all marked in-code:
  * wind read: wind is still the float atmosphere output (S2c) — step() reads float
    wind and quantizes the wind*dt_adv displacement + the d_eff (|wind|^2) coeff to
    Q16.16 at the boundary. So S2b is self-reproducible same-machine; cross-GPU
    once S2c. (// FLOAT BRIDGE until S2c)
  * fire smoke emission (fire_simulation.cpp): the emission delta smoke_emission*dt
    *I is quantized to Q16.16 and integer-added (order-free). Smoke param int32.
  * W5 steam puff (physics_engine.cpp step_water): the steam yield is quantized and
    integer-added into the (now int32) steam gas plane.
  * raycaster gas optics (render-only): gmap.gas is dequantized to a float scratch
    in game_renderer.compute_light_field + the physics_runner fire-light heat cast
    (gases never attenuate heat, so the heat deposit is unchanged) + the smoke
    overlay. One source of truth = the int field.

PYTHON: gamemap allocates gas (and the smoke view) as int32; new gas_fixed.py
quantize/dequantize helpers (mirror wave_fixed); field_edit "gas" dtype (smoke
policy — combine in real density, store quantized); recorder dequantizes smoke at
its debug boundary; input_handler debug-spawn writes FP_ONE counts; bindings take
int32 for smoke/gas across SmokeDynamics.step/sink_hop, FireSimulation.step,
run_substeps, step_water, step_tail.

GATES:
  * P1 bit-determinism GREEN (field_ab_harness: gas int32 + bit-identical run-to-
    run; the smoke seed quantized so the cloud actually transports). Added the S2b
    assert block (dtype + per-tick bit-identity + the deterministic-non-conservation
    mass-trace identity).
  * FEEL-REGRESSION vs the float SL (the centerpiece visual gate, _s2b_feel_
    regression.py — drives the ACTUAL C++ engine vs a faithful float SL on the real
    unhcr_vessel_2 ship map, the prototype's methodology): mean|d|=0.00032 (calm) /
    0.00038 (blast), max|d|=0.0037/0.0024 over the open domain — the filled,
    internally-structured SL look survives (float densities up to ~0.8; ~0.03% mean
    error, only faint edge-speckle). The look is preserved.
  * DETERMINISTIC NON-CONSERVATION (calm vs blast): keeps 58.3% of peak in calm,
    91.9% in a blast — matching the prototype's ~50%/~85%; bounded, gently decaying,
    never amplifying.
  * no-float ratchet re-baselined (smoke_dynamics.cpp float 47->25 fp:fast 1->0
    double 1->15; fire_simulation double 0->2; physics_engine float 71->66 double
    24->25) — the transport core is integer; the rises are the documented FLOAT
    BRIDGES.
  * Full suite 355 passed (was 354 + the new reciprocal proof); both --auto exit 0
    (default unhcr_vessel + unhcr_vessel_2, 600 frames each, render-side bridge OK).
  * Smoke goldens regenerated at the representation level (the SLint matches the
    float SL closely — the look is preserved, the prototype proved it). The
    multigas behaviour-preservation test is now BIT-IDENTICAL (both paths run the
    identical integer-SL with d_smoke 0.1).

+++ cpp/src/bindings.cpp
@@ -123,7 +123,7 @@ PYBIND11_MODULE(breach_physics, m) {
         // Patch 2b: step() is WIND-ONLY (no sink_x/sink_y) — the breach sink-pull
         // moved to sink_hop() below. dt_scale is removed (smoke moves on real dt).
         .def("step", [](const SmokeDynamics& self,
-                        py::array_t<float> smoke,
+                        py::array_t<int32_t> smoke,        // S2b: Q16.16 int32
                         py::array_t<float> wind_x,
                         py::array_t<float> wind_y,
                         py::array_t<bool>  obstacles,
@@ -146,7 +146,7 @@ PYBIND11_MODULE(breach_physics, m) {
         // Patch 2b: ONE 1-cell BFS-gradient breach pull (the decoupled sink). No
         // dt — each call is exactly one hop; the engine runs it K× per tick.
         .def("sink_hop", [](const SmokeDynamics& self,
-                            py::array_t<float> smoke,
+                            py::array_t<int32_t> smoke,    // S2b: Q16.16 int32
                             py::array_t<float> sink_x,
                             py::array_t<float> sink_y,
                             py::array_t<bool>  obstacles,
@@ -190,7 +190,7 @@ PYBIND11_MODULE(breach_physics, m) {
         .def("step", [](const FireSimulation& self,
                         py::array_t<float> fire,
                         py::array_t<float> atmosphere,
-                        py::array_t<float> smoke,
+                        py::array_t<int32_t> smoke,        // S2b: Q16.16 int32
                         py::array_t<float> wall_hp,
                         py::array_t<int32_t> temperature,
                         py::array_t<float> wind_x,
@@ -541,7 +541,7 @@ PYBIND11_MODULE(breach_physics, m) {
                              // fire group
                              py::array_t<float> fire,
                              py::array_t<float> atmosphere,
-                             py::array_t<float> smoke,
+                             py::array_t<int32_t> smoke,         // S2b: Q16.16 int32
                              py::array_t<float> wall_hp,
                              py::array_t<int32_t> temperature,
                              py::array_t<float> wind_x,
@@ -619,7 +619,7 @@ PYBIND11_MODULE(breach_physics, m) {
                                 py::array_t<bool>  is_vacuum,
                                 py::array_t<float> dyn_permeability,
                                 py::array_t<float> dyn_wave_absorb,
-                                py::array_t<float> gas,
+                                py::array_t<int32_t> gas,           // S2b: Q16.16 int32
                                 py::array_t<float> gas_diffusion,
                                 py::array_t<float> sink_x,
                                 py::array_t<float> sink_y,
@@ -638,7 +638,7 @@ PYBIND11_MODULE(breach_physics, m) {
             // gas: (N, h, w) contiguous — pass the base pointer + N; run_substeps
             // strides by plane (h*w) internally. h/w come from the 2D fields above.
             auto gv = gas.mutable_unchecked<3>();
-            float* gas_ptr = gv.mutable_data(0, 0, 0);
+            int32_t* gas_ptr = gv.mutable_data(0, 0, 0);        // S2b: Q16.16 int32
             const int n_gases = static_cast<int>(gv.shape(0));
             // gas_diffusion: (N,) float32 — the per-gas base-diffusion column.
             auto gd = gas_diffusion.unchecked<1>();
@@ -682,7 +682,7 @@ PYBIND11_MODULE(breach_physics, m) {
                               py::array_t<float> atmosphere,       // float (FLOAT BRIDGE)
                               py::array_t<int32_t> wave_p,         // S2a: Q16.16 int32
                               py::array_t<bool>  solid,
-                              py::array_t<float> gas,              // float (FLOAT BRIDGE: steam)
+                              py::array_t<int32_t> gas,            // S2b: Q16.16 int32 (steam puff quantized)
                               py::array_t<int32_t> before,         // S1: Q16.16 int32 snapshot
                               py::array_t<float> dyn_permeability, // float (FLOAT BRIDGE: seal)
                               int steam_idx, float tilt_x, float tilt_y,
@@ -702,7 +702,7 @@ PYBIND11_MODULE(breach_physics, m) {
             // gas: (N, h, w) contiguous — pass the base pointer; step_water strides
             // by plane (h*w) internally to reach the steam slice (steam_idx).
             auto gv = gas.mutable_unchecked<3>();
-            float* gas_ptr = gv.mutable_data(0, 0, 0);
+            int32_t* gas_ptr = gv.mutable_data(0, 0, 0);        // S2b: Q16.16 int32
             self.step_water(
                 wd, vx, vy, fl, atm, wp, sol,
                 gas_ptr, bef, perm,
